main.py
# main.py
import sys
sys.path.append('/home/runner/workspace/.pythonlibs/lib/python3.11/site-packages')
import backtrader as bt
import ta
import nsepy
from SmartApi import SmartConnect
import xgboost
print("All core dependencies imported successfully!")

def main():
    print("Quant Trading Bot - Project Initialized!")
    # Entry point for core pipeline
    # TODO: Add argument parsing, config loading, etc.

if __name__ == "__main__":
    main()
