requirements.txt
backtrader
ta
nsepy
smartapi-python
xgboost
python-telegram-bot
logzero
websocket